moved all components to seperate directories and relinked them
