--- src/Components/ImageSlide.js
@@ -1,4 +1,0 @@
-import React from 'react'
-import './../App.css'
-
-  export default ImageSlide
